requirements.txt
visdom
fire
torchvision
torchnet
# git+https://github.com/pytorch/tnt.git@master
tqdm
ipdb
